// src/data/types/causa.ts
export interface Causa {
  IdCausa: number
  IdCausaAugusta: string | null
  IdOrg: number | null
  NombreOrganismo: string | null
  Prefijo: string | null
  Numero: string | null
  Sufijo: string | null
  Caratula: string | null
  Estado: string | null
  LocalidadJuzgado: string | null
  IdUnicoCausa: string | null
}